# server/openapi.yaml
openapi: 3.0.3
info:
  title: Metro Verify API
  description: >
    Digital Legal Metrology Verification & Lifecycle Management Platform (SIH26036).
    Covers Owner, Administrator, and LMO/GATC Officer endpoints, plus the AI chatbot.
    Endpoints marked "(stored procedure)" call a MySQL stored procedure for an atomic,
    multi-step operation rather than a plain Prisma query.
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development backend
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Certificate:
      type: object
      properties:
        certificateId: { type: string }
        applicationId: { type: string }
        instrumentId: { type: string }
        certificateNumber: { type: string, example: "IND/LM/TS/26/0147" }
        qrPayload: { type: string }
        issuedDate: { type: string, format: date }
        validUntil: { type: string, format: date }
...
security:
  - bearerAuth: []
paths:
...
  /api/inspections/{id}/start:
    post:
      tags: [LMO - Inspections]
      summary: Mark an inspection as started (status → IN_PROGRESS)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Inspection marked in progress }
  /api/inspections/{id}/results:
    post:
      tags: [LMO - Inspections]
      summary: Record measurements and observations
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                measurements: { type: object }
                observations: { type: string }
      responses:
        "200": { description: Results recorded }
  /api/inspections/{id}/photos:
    post:
      tags: [LMO - Inspections]
      summary: Attach a captured photo to the inspection/evidence record
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo: { type: string, format: binary }
      responses:
        "200": { description: Photo attached }
  /api/inspections/{id}/location:
    post:
      tags: [LMO - Inspections]
      summary: Attach real GPS coordinates and timestamp to the inspection
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                latitude: { type: number }
                longitude: { type: number }
                capturedAt: { type: string, format: date-time }
      responses:
        "200": { description: Location evidence attached }
  /api/inspections/{id}/complete:
    post:
      tags: [LMO - Inspections]
      summary: >
        Complete a verification with a PASS result (stored procedure — sp_CompleteVerification).
        Atomically marks the inspection COMPLETED, sets instrument status VERIFIED,
        generates the certificate, and updates the officer's counters.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Verification completed, certificate generated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Certificate" }
  /api/inspections/{id}/failure:
    post:
      tags: [LMO - Inspections]
      summary: >
        Complete a verification with a FAIL result (stored procedure — sp_CompleteVerification).
        Sets status to FAILED/REINSPECTION_REQUIRED. No certificate is generated.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Marked as failed, no certificate generated }
  /api/inspections/{id}/reinspection:
    post:
      tags: [LMO - Inspections]
      summary: Create a reinspection task for a previously failed instrument
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Reinspection task created }

  # ---------------- LMO: OFFLINE SYNC ----------------
...